fix: Fix typical_p value edge case causing tensor OOB errors
In TypicalLogitsWarper we need to ensure that last_ind is a valid index for the vocab size.

In low precision situations, it can end up that the cumulative mass covers the entire vocab even when it is < 1 (specifically observed for typical_p = 0.99 with bfloat16).

Also clean up patched version of TopPLogitsWarper now that the fix is in transformers.