Option to omit stop sequence from text output
Introduces an optional bool include_stop_sequence field in the stopping criteria section of the generate request API.

There is a new global configuration option to control what the default behaviour should be for requests that don't specify this. For now the default is to include the stop sequence text, for backwards compatibility (though this is applied in the launcher, the default at the router level is to omit them).

There's also a new stop_sequence field in the generate API response message, which will contain the matched stop sequence iff stop_reason == STOP_SEQUENCE.

In addition, when a matched stop sequence is included in the output, the text will now be trimmed to the end of the stop sequence rather than potentially including additional characters from the last generated token.