Never output EOS token's text
There is a global output-special-tokens option which is disabled by default, but when enabled causes all special tokens to be output including the terminating EOS token.

For the use cases we've encountered this is undesirable. I can't think of a case where this would be needed/wanted since it will only ever be at the end of the output, and in these cases the returned stop_reason will be EOS_TOKEN.