Python tests revival
This PR does a bit of tidying to bring back in the python tests.

This mainly kills off the broken tests from before the grand schism with TGI, which was mainly tests/models and things involving StoppingCriteria. After fixing a few other tests that seem to have been broken by recent changes, make python-tests is good to go again.