build: update Flash Attention v2 cache build and install
Makes some changes to the Flash Attention v2 build and install to simplify and to install as a complete python wheel package including metadata (not just copying a subset of the files). The metadata is needed so that transformers can detect the installation (transformers uses importlib.metadata.version() to inspect the metadata and parse a version).

NB: The format of the cache image is changed with this PR so older versions will no longer work.

Signed-off-by: Travis Johnson <tsjohnso@us.ibm.com>