feat: add support for models from ibm-fms
Initially this is focused on supporting Calico architecture models, including from `gpt_megatron` format checkpoints.

Note: Even just importing / registering FMS model classes causes the process to be forked (for a warmup for PyTorch Compile). To avoid that, we can inspect the model type listed in the config.json before doing the import of FMS.

Signed-off-by: Travis Johnson <tsjohnso@us.ibm.com>