First attempt at rocm build
As far as I could see, the pytoch wheel index doesn't have a package
with the latest rocm version and the latest AMD wheels aren't built
with the latest pytorch. For now I'm prioritizing the AMD wheel.

To make it work on my Ryzen 7 7700X I had to set the GFX version:
```
export HSA_OVERRIDE_GFX_VERSION=10.3.0
```
The correct version for your hardware can be found with `rocminfo`