🔥 Remove our exllama code because we use auto-gptq vendored kernels (#59)
We recently found that AutoGPTQ vendors its own versions of exllama and
exllamav2 kernels in
[augotgptq_extension](https://github.com/AutoGPTQ/AutoGPTQ/tree/main/autogptq_extension)
that are installed with the library. Since we install AutoGPTQ after we
installed our own builds of the exllama kernels, the AutoGPTQ ones
overwrite our copies. So it turns out that we don't need to vendor and
compile our own exllama kernels.

Signed-off-by: Travis Johnson <tsjohnso@us.ibm.com>