Fix division by zero errors
When the memory characterizer doesn't find a linear or quadratic
behavior, the coefficients are set to zero resulting in division by zero
errors in the inverse functions.

In this commit this situation is detected and the max float is returned
to be consistent with the semantics of the memory estimator.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>