Free blocks in KVCacheManager upon error (#96)
#### Motivation

We are see pods with spec. decoding getting restarted in BAM due to
health checks failing. Upon inspection of the logs, it looks like we are
running out of blocks, and never recovering from it.
#### Modifications

I added a simple check that if something goes wrong when generating a
token, we free the blocks associated with that batch. I also had to
ensure that the we free the child sequences that get created during
speculation if something goes wrong there too.

#### Result

I've verified this allow us to recover from failures related to running
out of blocks. Hopefully after this fix, we don't see the inference
server getting restarted.

Signed-off-by: Thomas Parnell <tpa@zurich.ibm.com>